requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Auth / security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client
httpx>=0.27.0
requests>=2.31.0

# Data
pandas>=2.0.0
yfinance>=0.2.36
python-bcb>=0.3.0

# Quantitative analysis
quantstats>=0.0.620

# Visualization
matplotlib==3.8.2
seaborn>=0.12.0

# AI / LLM
groq>=0.4.0

# MCP server
mcp>=1.0.0

# Polymarket trading
py-clob-client>=0.17,<1.0
web3>=6.0,<7.0
eth-account>=0.11,<1.0

# Scraping
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Brazilian market data
mercados>=0.2.0

# Environment
python-dotenv>=1.0.0
